refactor: create pkg
