+++ cmd/main.go
@@ -3,11 +3,13 @@ package main
 import (
 	"flag"
 	"fmt"
+
+	gotoPkg "github.com/gabrielg2020/goto/cmd/pkg"
 )
 
 func main() {
 	// Get the configuration
-	config, err := LoadConfig()
+	config, err := gotoPkg.LoadConfig()
 	if err != nil {
 		fmt.Printf("Error: %v\n", err)
 		return
@@ -26,7 +28,7 @@ func main() {
 	pattern := args[0]
 
 	// Call the search and select function
-	results, err := SearchAndSelect(".", *maxDepth, pattern, config.ExcludeDirs)
+	results, err := gotoPkg.SearchAndSelect(".", *maxDepth, pattern, config.ExcludeDirs)
 	if err != nil {
 		fmt.Printf("Error: %v\n", err)
 		return
